requirements.txt
streamlit
tavily-python
openai
requests